requirements.txt
websocket-client
yfinance
starlette
bcrypt
python-jose
fastapi-utils
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
aiohttp
websockets
python-dotenv
PyQt6
requests
asyncio
passlib[bcrypt]